pyproject.toml
[project]
name = "sps-genai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pillow>=12.0.0",
    "pydantic>=2.11.7",
    "scikit-learn>=1.7.2",
    "spacy>=3.8.7",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "uvicorn[standard]>=0.35.0",
]
